fix: pulish entities

+++ backend/src/main/java/ar/net/ut/backend/user/entity/UserProfile.java
@@ -37,7 +37,7 @@ public class UserProfile {
     @Column(columnDefinition = "jsonb")
     private Map<Preference, String> preferences;
 
-    private int averageContributionPoints;
+    private double averageContributionPoints;
 
     public Map<Preference, String> getPreferences() {
         return preferences != null ? Collections.unmodifiableMap(preferences) : Collections.emptyMap();
